# By Christoph Schwarzbach (4) and Patrick Belliveau (4)
# Via Patrick Belliveau (3) and GitHub (1)
* 'master' of https://github.com/JuliaInv/JOcTree:
  Finished updating for 0.5 and improved efficiency of edge and face mass matrix generation. getFaceMassMatrix in particular was very slow. now almost 8 times faster than previously.
  Fixed bug in regularizeOcTree2.jl. SparseMatrices are now immutable. regularizeOcTree2 was editing sparse matrix fields directly.
  Updated for Julia version 0.5 and Julia 0.5 compatible version of jInv
  Added new uniteOcTrees method that reduces an array of OcTrees to a single OcTree.
  Make getEdgeIntegralOfPolygonalChain faster
  Improve speed of getEdgeIntegralOfPolygonalChain
  Remove restriction that polygon must be contained in fine cell region of OcTree
  Add optional parameter to createOcTreeFromBox and tighten bounding box

Conflicts:
	src/OcTreeMeshFEM.jl
	src/OcTreeMeshFV.jl
	src/getNodalConstraints.jl
	src/sparse3.jl

+++ src/sparse3.jl
@@ -8,8 +8,8 @@ import Base.setindex!
 import Base.ndims 
 
 type SparseArray3D
-        SV::SparseMatrixCSC{Int}
-        sz::Vector{Int}    # size of fine mesh
+        SV::SparseMatrixCSC{Int64}
+        sz::Vector{Int64}    # size of fine mesh
 end
 
 Base.copy(S::SparseArray3D) = SparseArray3D(copy(S.SV),copy(S.sz))
